+++ .eslintrc.js
@@ -2,6 +2,9 @@ const path = require('path');
 
 module.exports = {
   extends: ['hzdg', 'hzdg/jest', 'hzdg/react', 'hzdg/typescript'],
+  globals: {
+    __DEV__: 'readonly',
+  },
   env: {
     node: true,
     es6: true,