+++ babel.config.js
@@ -26,6 +26,7 @@ module.exports = {
     '@babel/preset-react',
   ],
   plugins: [
+    'babel-plugin-dev-expression',
     '@babel/plugin-proposal-class-properties',
     '@babel/plugin-proposal-nullish-coalescing-operator',
     '@babel/plugin-proposal-object-rest-spread',