+++ packages/behaviors/hzcore-gesture-catcher/src/examples/TouchGestureVisualizer.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import {useTouchGesture} from '@hzcore/gesture-catcher';
-import GestureVisualizer, {
-  useGestureVisualizer,
-} from './components/GestureVisualizer';
+import GestureVisualizer, {useGestureVisualizer} from 'gesture-visualizer';
 import GestureConfigurator, {
   useGestureConfigurator,
 } from './components/GestureConfigurator';