+++ packages/utilities/hzcore-moving-average/src/index.ts
@@ -41,10 +41,11 @@ export interface MovingAverageOptions {
  */
 function weigh(value: number, max: number, scale = 0): number {
   let weight = 1;
+  if (!scale) return weight;
   if (scale > 0) {
     weight = ((max - value) / max) * scale;
   } else if (scale < 0) {
-    weight = (value / max) * Math.abs(scale);
+    weight = ((value + 1) / max) * Math.abs(scale);
   }
   return weight;
 }
@@ -78,9 +79,56 @@ export default class MovingAverage {
     this._average = null;
   }
 
+  /**
+   * Returns a portion of the stored values in this `MovingAverage`
+   * as a new `MovingAverage`. The new `MovingAverage` will inherit the
+   * configuration of this `MovingAverage`, with the exception of `size`.
+   * The size of the new `MovingAverage` will be determined by `end - start`,
+   * where `end` defaults to the size of the originating `MovingAverage`,
+   * and `start` defaults to `0`.
+   */
+  slice(start?: number, end?: number): MovingAverage {
+    start = start ?? 0;
+    if (start > this._size) start = this._size;
+    if (start < 0) start = Math.max(0, this._size + start);
+
+    end = end ?? this._size;
+    if (end > this._size) end = this._size;
+    if (end < 0) end = Math.max(start, this._size + end);
+    if (end < start) end = start;
+
+    const size = end - start;
+    if (!size) {
+      return new MovingAverage({
+        size: 1,
+        weight: this._scale,
+        round: this._round,
+      });
+    }
+
+    const sliced = new MovingAverage({
+      size,
+      weight: this._scale,
+      round: this._round,
+    });
+    const pCount = Math.min(this._count, this._size);
+    const pStart = (this._pointer + 1 - pCount + pCount) % pCount;
+    const count = Math.min(size, pCount);
+    let i = 0;
+    while (i < count) {
+      const index = (pStart + start + i) % pCount;
+      const value = this._store[index];
+      sliced.push(value);
+      i++;
+    }
+    return sliced;
+  }
+
   /** Add a value to the moving average. */
   push(v: number): void {
-    this._pointer = (this._pointer + 1) % this._size;
+    this._pointer = this._count
+      ? (this._pointer + 1) % this._size
+      : this._pointer;
     this._store[this._pointer] = v;
     this._average = null;
     this._delta += v;
@@ -89,6 +137,7 @@ export default class MovingAverage {
 
   /** View the last added value. */
   peek(): number {
+    let count = Math.min(this._count, this._size);
     return this._store[this._pointer] ?? NaN;
   }
 
@@ -132,12 +181,14 @@ export default class MovingAverage {
     if (this._average === null) {
       let sum = 0;
       let sumWeight = 0;
+      let count = Math.min(this._count, this._size);
       let i = 0;
-      while (i < this._count) {
-        const index = (this._pointer + i) % this._size;
+
+      while (i < count) {
+        const index = (this._pointer - i + count) % count;
         const value = this._store[index];
         if (value != null) {
-          const weight = weigh(i, this._count, this._scale);
+          const weight = weigh(i, count, this._scale);
           sumWeight += weight;
           sum += value * weight;
         }