+++ packages/behaviors/hzcore-page-gesture/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.12](https://github.com/hzdg/hz-core/compare/@hzcore/page-gesture@2.0.11...@hzcore/page-gesture@2.0.12) (2020-02-26)
+
+**Note:** Version bump only for package @hzcore/page-gesture
+
+
+
+
+
+## [2.0.11](https://github.com/hzdg/hz-core/compare/@hzcore/page-gesture@2.0.10...@hzcore/page-gesture@2.0.11) (2020-02-13)
+
+**Note:** Version bump only for package @hzcore/page-gesture
+
+
+
+
+
+## [2.0.10](https://github.com/hzdg/hz-core/compare/@hzcore/page-gesture@2.0.9...@hzcore/page-gesture@2.0.10) (2020-02-13)
+
+**Note:** Version bump only for package @hzcore/page-gesture
+
+
+
+
+
+## [2.0.9](https://github.com/hzdg/hz-core/compare/@hzcore/page-gesture@2.0.8...@hzcore/page-gesture@2.0.9) (2020-02-13)
+
+**Note:** Version bump only for package @hzcore/page-gesture
+
+
+
+
+
 ## [2.0.8](https://github.com/hzdg/hz-core/compare/@hzcore/page-gesture@2.0.7...@hzcore/page-gesture@2.0.8) (2020-02-10)